+++ packages/components/icon-button/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/icon-button
 
+## 0.0.6
+
+### Patch Changes
+
+- fix disable opacity
+
 ## 0.0.5
 
 ### Patch Changes