+++ packages/components/stepper/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/stepper
 
+## 0.0.14
+
+### Patch Changes
+
+- fix disable opacity
+
 ## 0.0.13
 
 ### Patch Changes