+++ packages/react/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @jamsr-ui/react
 
+## 0.0.207
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/autocomplete@0.0.109
+  - @jamsr-ui/icon-button@0.0.6
+  - @jamsr-ui/typography@0.0.81
+  - @jamsr-ui/checkbox@0.0.76
+  - @jamsr-ui/stepper@0.0.14
+  - @jamsr-ui/avatar@0.0.77
+  - @jamsr-ui/button@0.0.80
+  - @jamsr-ui/rating@0.0.36
+  - @jamsr-ui/select@0.0.98
+  - @jamsr-ui/switch@0.0.85
+  - @jamsr-ui/input@0.0.102
+  - @jamsr-ui/radio@0.0.83
+  - @jamsr-ui/card@0.0.86
+  - @jamsr-ui/menu@0.0.97
+  - @jamsr-ui/theme@0.0.75
+  - @jamsr-ui/rhf@0.0.52
+  - @jamsr-ui/data-table@0.0.67
+  - @jamsr-ui/dialog@0.0.91
+  - @jamsr-ui/drawer@1.0.77
+  - @jamsr-ui/editor@0.0.108
+  - @jamsr-ui/popover@0.0.77
+  - @jamsr-ui/tooltip@0.0.73
+  - @jamsr-ui/alert@0.0.92
+  - @jamsr-ui/confirmation@0.0.72
+  - @jamsr-ui/accordion@0.0.84
+  - @jamsr-ui/tags-input@0.0.39
+  - @jamsr-ui/chip@0.0.87
+  - @jamsr-ui/header@0.0.80
+  - @jamsr-ui/tab@0.0.96
+  - @jamsr-ui/textarea@0.0.102
+  - @jamsr-ui/toast@0.0.34
+
 ## 0.0.206
 
 ### Patch Changes