+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/theme
 
+## 0.0.75
+
+### Patch Changes
+
+- fix disable opacity
+
 ## 0.0.74
 
 ### Patch Changes